+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Mage class, magic attacks and ether item #78
 * Quests to raise 5 levels on each available player class #81
 * Reach level 50 and level 100 quests #81
+* Items rewarded on battle won #82
 
 ### Removed
 * Backwards compatibility code for binary game data from v0.4.0 #75